+++ src/QuickMarcEditor/index.js
@@ -3,3 +3,4 @@ export { default as QuickMarcEditorContainer } from './QuickMarcEditorContainer'
 export { default as QuickMarcDeriveWrapper } from './QuickMarcDeriveWrapper';
 export { default as QuickMarcCreateWrapper } from './QuickMarcCreateWrapper';
 export { default as QuickMarcEditWrapper } from './QuickMarcEditWrapper';
+export { default as QuickMarcCreateBibWrapper } from './QuickMarcCreateBibWrapper';